+++ test/c99/test-errors.c
@@ -2685,6 +2685,17 @@ static int test_site_gcrs_posvel() {
   return n;
 }
 
+static int test_timescale_diff() {
+  int n = 0;
+
+  novas_timespec ts = {};
+
+  if(check_nan("timescale_diff:ts:null", novas_timescale_offset(NULL, NOVAS_TAI, NOVAS_TT))) n++;
+  if(check_nan("timescale_diff:scale:-1", novas_timescale_offset(&ts, (enum novas_timescale) -1, NOVAS_TT))) n++;
+  if(check_nan("timescale_diff:ref:-1", novas_timescale_offset(&ts, NOVAS_TT, (enum novas_timescale) -1))) n++;
+
+  return n;
+}
 
 int main(int argc, const char *argv[]) {
   int n = 0;
@@ -2916,6 +2927,7 @@ int main(int argc, const char *argv[]) {
   if(test_uvw()) n++;
   if(test_site_uvw()) n++;
   if(test_site_gcrs_posvel()) n++;
+  if(test_timescale_diff()) n++;
 
   if(n) fprintf(stderr, " -- FAILED %d tests\n", n);
   else fprintf(stderr, " -- OK\n");